Add appsignal.demo task
As currently present in our Ruby gem as well:
https://github.com/appsignal/appsignal-ruby/pull/199

The "appsignal.demo" command is useful for debugging AppSignal and
testing AppSignal.

It sends a demonstration sample for an error and a performance issue.
They will show up on AppSignal.com as demonstration samples with a small
special UI attached to them describing their purpose.

Debugging manual referencing the "demo" command:
http://docs.appsignal.com/support/debugging.html